test(endpoint-syndicate): use node test runner
