+++ packages/endpoint-syndicate/test/integration/200-syndicates-url-for-target.js
@@ -1,0 +1,42 @@
+import { strict as assert } from "node:assert";
+import { after, before, describe, it } from "node:test";
+import supertest from "supertest";
+import { mockAgent } from "@indiekit-test/mock-agent";
+import { testServer } from "@indiekit-test/server";
+import { testToken } from "@indiekit-test/token";
+
+await mockAgent("endpoint-syndicate");
+const server = await testServer({
+  plugins: ["@indiekit/syndicator-mastodon"],
+});
+const request = supertest.agent(server);
+
+describe("endpoint-syndicate POST /syndicate", () => {
+  before(async () => {
+    await request
+      .post("/micropub")
+      .auth(testToken(), { type: "bearer" })
+      .set("accept", "application/json")
+      .send("h=entry")
+      .send("name=foobar")
+      .send("mp-syndicate-to=https://mastodon.example/@username");
+  });
+
+  it("Syndicates a URL", async () => {
+    const result = await request
+      .post("/syndicate")
+      .set("accept", "application/json")
+      .query({ source_url: "https://website.example/notes/foobar/" })
+      .query({ token: testToken() });
+
+    assert.equal(result.status, 200);
+    assert.equal(
+      result.body.success_description,
+      "Post updated at https://website.example/notes/foobar/",
+    );
+  });
+
+  after(() => {
+    server.close(() => process.exit(0));
+  });
+});
