+++ packages/endpoint-syndicate/test/integration/403-syndicates-url-missing-update-scope.js
@@ -1,0 +1,48 @@
+import { strict as assert } from "node:assert";
+import { after, before, describe, it } from "node:test";
+import supertest from "supertest";
+import { mockAgent } from "@indiekit-test/mock-agent";
+import { testServer } from "@indiekit-test/server";
+import { testToken } from "@indiekit-test/token";
+
+await mockAgent("endpoint-syndicate");
+const server = await testServer({
+  plugins: ["@indiekit/syndicator-mastodon"],
+});
+const request = supertest.agent(server);
+const token = testToken({ scope: "create" });
+
+describe("endpoint-syndicate POST /syndicate", () => {
+  before(async () => {
+    await request
+      .post("/micropub")
+      .auth(token, { type: "bearer" })
+      .set("accept", "application/json")
+      .send("h=entry")
+      .send("name=foobar")
+      .send("mp-syndicate-to=https://mastodon.example/@username");
+  });
+
+  it("Syndicates a URL and redirects", async () => {
+    const result = await request
+      .post("/syndicate")
+      .set("accept", "application/json")
+      .send({
+        syndication: {
+          url: "https://website.example/notes/foobar/",
+          redirect_uri: "/posts/12345",
+        },
+        token,
+      });
+
+    assert.equal(result.status, 403);
+    assert.equal(
+      result.body.error_description,
+      "The request requires higher privileges than provided by the access token",
+    );
+  });
+
+  after(() => {
+    server.close(() => process.exit(0));
+  });
+});
