--- packages/endpoint-syndicate/tests/integration/200-syndicates-url-for-target.js
@@ -1,34 +1,0 @@
-import test from "ava";
-import supertest from "supertest";
-import { mockAgent } from "@indiekit-test/mock-agent";
-import { testServer } from "@indiekit-test/server";
-import { testToken } from "@indiekit-test/token";
-
-await mockAgent("endpoint-syndicate");
-
-test("Syndicates a URL", async (t) => {
-  const server = await testServer({
-    plugins: ["@indiekit/syndicator-mastodon"],
-  });
-  const request = supertest.agent(server);
-  await request
-    .post("/micropub")
-    .auth(testToken(), { type: "bearer" })
-    .set("accept", "application/json")
-    .send("h=entry")
-    .send("name=foobar")
-    .send("mp-syndicate-to=https://mastodon.example/@username");
-  const result = await request
-    .post("/syndicate")
-    .set("accept", "application/json")
-    .query({ source_url: "https://website.example/notes/foobar/" })
-    .query({ token: testToken() });
-
-  t.is(result.status, 200);
-  t.is(
-    result.body.success_description,
-    "Post updated at https://website.example/notes/foobar/",
-  );
-
-  server.close(t);
-});
